backend/py-service/requirements.txt
# Requirements for Python 3.10
# AWS
# Cache
# Configuration
# Database
# Development
# FastAPI and ASGI
# Security
# Utilities
alembic==1.13.1
argon2-cffi==23.1.0
asyncpg==0.29.0
black==24.1.1
boto3==1.34.20
cryptography==42.0.2
email-validator>=2.2.0
fastapi==0.109.0
hiredis==2.3.2
httpx==0.26.0
mypy==1.8.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
pydantic==2.5.3
pytest-asyncio==0.23.3
pytest==7.4.4
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==5.0.1
ruff==0.1.15
sqlalchemy[asyncio]==2.0.25
sse-starlette==1.8.2
uvicorn[standard]==0.27.0
aiosmtplib==3.0.1
gunicorn==21.2.0
